requirements.txt
customtkinter~=5.2.2
pillow~=12.0.0
mysql-connector-python~=9.5.0